requirements.txt
# requirements.txt - AI Chat dependencies
# Core
flask>=3.0.0
python-dotenv>=1.0.0
requests>=2.31.0

# Google Gemini AI
google-genai>=1.0.0

# Data processing
pydantic>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Document processing
python-docx>=1.0.0

# Production server
gunicorn>=21.0.0
